Code review updates.
